change sql

+++ docker/initsql/a-n9e.sql
@@ -204,7 +204,7 @@ CREATE TABLE `alert_rule` (
     `prod` varchar(255) not null default '',
     `algorithm` varchar(255) not null default '',
     `algo_params` varchar(255),
-    `delay` tinyint(1) not null default 0,
+    `delay` int not null default 0,
     `severity` tinyint(1) not null comment '0:Emergency 1:Warning 2:Notice',
     `disabled` tinyint(1) not null comment '0:enabled 1:disabled',
     `prom_for_duration` int not null comment 'prometheus for, unit:s',
